% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{bernstein_copula_pdf}
\alias{bernstein_copula_pdf}
\title{Bernstein Copula PDF}
\usage{
bernstein_copula_pdf(u, v, R)
}
\arguments{
\item{u}{A numeric value in [0, 1], the first copula argument.}

\item{v}{A numeric value in [0, 1], the second copula argument.}

\item{R}{A numeric matrix of size m x n representing the copula weights.}
}
\value{
A numeric value giving the copula density at (u, v).
}
\description{
Computes the probability density function of the Bernstein copula at a given point (u, v).
}
\examples{
R <- matrix(1/4, 2, 2)
bernstein_copula_pdf(0.5, 0.5, R)

}
