% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-group.R
\name{bincount2d}
\alias{bincount2d}
\title{Bin counts for 2D data}
\usage{
bincount2d(x, y, xbreaks = "Sturges", ybreaks = "Sturges")
}
\arguments{
\item{x}{Numeric vector. First variable to be binned.}

\item{y}{Numeric vector. Second variable to be binned.}

\item{xbreaks}{Numeric or character. Breaks for the x variable. Default is "Sturges".}

\item{ybreaks}{Numeric or character. Breaks for the y variable. Default is "Sturges".}
}
\value{
A list containing:
\item{xbreaks}{Breaks for the x variable.}
\item{ybreaks}{Breaks for the y variable.}
\item{counts}{A matrix of counts for each bin.}
The dimensions of the matrix correspond to the number of bins in x and y.
The row names correspond to the x bins and the column names correspond to the y bins.
}
\description{
This function takes two numeric vectors and bins them into a 2D histogram.
It returns a matrix of counts for each bin.
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
counts <- bincount2d(x, y)
print(counts)
}
