% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{empdist}
\alias{empdist}
\title{Empirical Marginal Distribution Constructor}
\usage{
empdist(data, ecdf = NULL, density = NULL)
}
\arguments{
\item{data}{Numeric vector. Data points for the empirical distribution.}

\item{ecdf}{Function. Empirical CDF function (default is \code{ecdf(data)}).}

\item{density}{Function. Density function (default is \code{density(data)}).}
}
\value{
A \code{MarginalDistribution} R6 object with fields \code{params}, \code{pdf}, \code{cdf}, \code{ccdf}, and \code{emstep_func}.
}
\description{
Creates a \code{MarginalDistribution} object for an empirical distribution
based on the provided data.
}
