% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-exponential.R
\name{exponentialdist}
\alias{exponentialdist}
\title{Exponential Marginal Distribution Constructor}
\usage{
exponentialdist(rate = 1)
}
\arguments{
\item{rate}{Numeric. Rate parameter (lambda) of the exponential distribution (default is 1).}
}
\value{
An object of class \code{MarginalDistribution} with appropriate \code{pdf}, \code{cdf}, and \code{ccdf} methods.
}
\description{
Creates a \code{MarginalDistribution} object for an exponential distribution
with a specified rate parameter.
}
\examples{
fx <- exponentialdist(rate = 2)
fx$pdf(1)
fx$cdf(1)
fx$ccdf(1)

}
