% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-normal.R
\name{normaldist}
\alias{normaldist}
\title{Normal Marginal Distribution Constructor}
\usage{
normaldist(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Numeric. Mean of the normal distribution (default is 0).}

\item{sd}{Numeric. Standard deviation of the normal distribution (default is 1).}
}
\value{
A \code{MarginalDistribution} R6 object with fields \code{params}, \code{pdf}, \code{cdf}, \code{ccdf}, and \code{emstep_func}.
}
\description{
Creates a \code{MarginalDistribution} object for a normal (Gaussian) distribution,
with specified mean and standard deviation.
}
