% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_ph.R
\name{phdist}
\alias{phdist}
\title{Phase-type marginal distribution constructor.}
\usage{
phdist(alpha = c(0.5, 0.5), rate = c(1, 1))
}
\arguments{
\item{alpha}{A numeric vector of length k, representing the initial probabilities of the k states.}

\item{rate}{A numeric vector of length k, representing the rates of the k states.}
}
\value{
An object of class MarginalDistribution with appropriate pdf, cdf, and ccdf methods.
}
\description{
Creates a MarginalDistribution object for a phase-type distribution.
}
\examples{
fx <- phdist(alpha = c(0.5, 0.5), rate = c(1, 2))
fx$pdf(1)
fx$cdf(1)
fx$ccdf(1)

}
