% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{normaldist}
\alias{normaldist}
\title{Normal Marginal Distribution Constructor}
\usage{
normaldist(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Numeric. Mean of the normal distribution (default is 0).}

\item{sd}{Numeric. Standard deviation of the normal distribution (default is 1).}
}
\value{
An object of class \code{MarginalDistribution} with appropriate \code{pdf}, \code{cdf}, and \code{ccdf} methods.
}
\description{
Creates a \code{MarginalDistribution} object for a normal (Gaussian) distribution,
with specified mean and standard deviation.
}
\examples{
fx <- normaldist(mean = 0, sd = 1)
fx$pdf(0)
fx$cdf(0)
fx$ccdf(0)

}
